pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools>=68" ]

[project]
name = "deeplabcut-live-gui"
version = "2.0.0rc1"
description = "PySide6-based GUI to run real time pose estimation experiments with DeepLabCut"
readme = "README.md"
keywords = [ "deep learning", "deeplabcut", "gui", "pose estimation", "real-time" ]
license-files = [ "LICENSE" ]
requires-python = ">=3.10"  # don' forget to update tool.pyproject-fmt if this changes
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: X11 Applications :: Qt",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "cv2-enumerate-cameras",
  "deeplabcut-live==1.1",
  "matplotlib",
  "numpy",
  "opencv-python",
  "pydantic>=2",
  "pyside6",
  "qdarkstyle",
  "vidgear[core]",
]
[[project.authors]]
name = "M-Lab of Adaptive Intelligence"
email = "mackenzie@deeplabcut.org"
[[project.authors]]
name = "Mathis Group for Computational Neuroscience and AI"
email = "alexander@deeplabcut.org"
[project.optional-dependencies]
all = [ "harvesters", "pypylon" ]
basler = [ "pypylon" ]
dev = [
  "hypothesis>=6",
  "pre-commit",
  "pytest>=7",
  "pytest-cov>=4",
  "pytest-mock>=3.10",
  "pytest-qt>=4.2",
]
gentl = [ "harvesters" ]
pytorch = [
  "deeplabcut-live[pytorch]==1.1", # includes timm and scipy
]
test = [
  "hypothesis>=6",
  "pytest>=7",
  "pytest-cov>=4",
  "pytest-mock>=3.10",
  "pytest-qt>=4.2",
  "tox",
  "tox-gh-actions",
]
tf = [
  "deeplabcut-live[tf]==1.1",
]
[project.scripts]
dlclivegui = "dlclivegui:main"
[project.urls]
"Bug Tracker" = "https://github.com/DeepLabCut/DeepLabCut-live-GUI/issues"
Documentation = "https://github.com/DeepLabCut/DeepLabCut-live-GUI"  # FIXME @C-Achard replace once docs are up
Homepage = "https://github.com/DeepLabCut/DeepLabCut-live-GUI"
Repository = "https://github.com/DeepLabCut/DeepLabCut-live-GUI"

# [tool.setuptools]
# include-package-data = true
[tool.setuptools.package-data]
"dlclivegui.assets" = [ "*.png" ]
[tool.setuptools.packages]
find.where = [ "." ]
find.include = [ "dlclivegui*" ]
find.exclude = [ "tests*", "docs*" ]

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true
[tool.ruff.lint]
select = [ "B", "E", "F", "I", "UP" ]
ignore = [ "E741" ]

[tool.pyproject-fmt]
max_supported_python = "3.12"
generate_python_version_classifiers = true
# Avoid collapsing tables to field.key = value format (less readable)
table_format = "long"

[tool.pytest]
[tool.pytest.ini_options]
testpaths = [ "tests" ]
python_files = [ "test_*.py" ]
python_classes = [ "Test*" ]
python_functions = [ "test_*" ]
addopts = [
  "--strict-markers",
  "--strict-config",
  "--disable-warnings",
  # "--maxfail=1",
  "-ra",
  "-q",
]
markers = [
  "unit: Unit tests for individual components",
  "integration: Integration tests for component interaction",
  "functional: Functional tests for end-to-end workflows",
  "hardware: Tests that require specific hardware, notable camera backends",
  # "slow: Tests that take a long time to run",
  "gui: Tests that require GUI interaction",
  "timeout: Test timeout in seconds (pytest-timeout)",
]

[tool.coverage]
[tool.coverage.report]
exclude_lines = [
  "@abstract",
  "def __repr__",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "pragma: no cover",
  "raise AssertionError",
  "raise NotImplementedError",
]
omit = [
  "tests/*",
]
[tool.coverage.run]
branch = true
omit = [
  "*/__pycache__/*",
  # "*/site-packages/*",
  # breaks CI coverage reporting as it excludes our own installed package
]
source = [ "dlclivegui", "tests" ]
